requirements.txt
playwright
tf-playwright-stealth
fake-useragent
requests
requests[socks]